Move commands to their respective folders
