+++ src/main/java/frc/robot/commands/climber/ClimbLidar.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.climber;
 
 public class ClimbLidar {
